Kinda fixed elements in places, but its monkeypatch
