+++ src/components/CellComponentsGroup/_cell.scss
@@ -6,8 +6,16 @@
   border-radius: 10px;
   border: 2px solid rgba(var(--color-background-alter), 1);
 
-  width: calc(100% - 20px - 4px);
-  aspect-ratio: 1 / 1;
+  &-height-1 {
+    /* понятия не имею почему .37 */
+    aspect-ratio: 2.37/1;
+  }
+  &-height-2 {
+    aspect-ratio: 1/1;
+  }
+  &-height-3 {
+    aspect-ratio: 1/3;
+  }
 
   /* странные ширина и высота без установки этого */
   & > * {
@@ -16,6 +24,8 @@
   }
 
   &-grid {
+    max-width: 800px;
+    height: 100%;
     width: 100%;
 
     padding: 0;
@@ -23,9 +33,9 @@
     list-style-type: none;
 
     display: grid;
-    // grid-gap: 15px;
     row-gap: 15px;
     column-gap: 15px;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
+    grid-template-columns: repeat(4, 1fr);
+    grid-template-rows: 1fr;
   }
 }
